Move buffer types to common
