+++ src/SelectionManager.test.ts
@@ -7,7 +7,8 @@ import { assert } from 'chai';
 import { SelectionManager, SelectionMode } from './SelectionManager';
 import { SelectionModel } from './SelectionModel';
 import { BufferSet } from './BufferSet';
-import { ITerminal, IBuffer } from './Types';
+import { ITerminal } from './Types';
+import { IBuffer } from 'common/buffer/Types';
 import { IBufferLine } from 'common/Types';
 import { MockTerminal, MockCharSizeService, MockOptionsService, MockBufferService } from './TestUtils.test';
 import { BufferLine, CellData } from 'common/buffer/BufferLine';
