+++ src/public/Terminal.ts
@@ -4,8 +4,9 @@
  */
 
 import { Terminal as ITerminalApi, ITerminalOptions, IMarker, IDisposable, ILinkMatcherOptions, ITheme, ILocalizableStrings, ITerminalAddon, ISelectionPosition, IBuffer as IBufferApi, IBufferLine as IBufferLineApi, IBufferCell as IBufferCellApi } from 'xterm';
-import { ITerminal, IBuffer } from '../Types';
+import { ITerminal } from '../Types';
 import { IBufferLine } from 'common/Types';
+import { IBuffer } from 'common/buffer/Types';
 import { Terminal as TerminalCore } from '../Terminal';
 import * as Strings from '../Strings';
 import { IEvent } from 'common/EventEmitter2';
